tests/wpt/web-platform-tests/html/dom/render-blocking/header-inserted-preload-link.tentative.html

<!doctype html>
<title>Header-inserted preload links with "blocking=render" are render-blocking</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/preload/resources/preload_helper.js"></script>
<script src="support/test-render-blocking.js"></script>

<style>
@font-face {
  font-family: custom-font;
  src: url('/fonts/Ahem.ttf?pipe=trickle(d1)');
}
</style>
<span id="target" style="font: 20px/1 custom-font">Lorem ipsum</span>

<script>
test_render_blocking(
    '/fonts/Ahem.ttf?pipe=trickle(d1)',
    () => {
      const target = document.getElementById('target');
      assert_equals(target.offsetHeight, 20);
      assert_equals(target.offsetWidth, 220);
    },
    'Render-blocking web font is applied');
</script>
